requirements.txt
psycopg2-binary==2.9.10
pandas==2.3.2
python-dotenv==1.1.1
